Move function_type.py into numba/experimental
