+++ numba/experimental/function_type.py
